// admin/types/database.ts
export type SubscriptionStatus = 'trialing' | 'active' | 'canceled' | 'expired'
export type SubscriptionPlan = 'monthly' | 'pack_3m'
export type PaymentMethod = 'transfer' | 'card'
export type Currency = 'ARS' | 'USD'